+++ spec/tbank_grpc/client_spec.rb
@@ -36,9 +36,24 @@ RSpec.describe TbankGrpc::Client do
     expect(client.instruments).to equal(client.instruments)
   end
 
+  it 'exposes #users as UsersService' do
+    channel = GRPC::Core::Channel.new('localhost:50051', {}, :this_channel_is_insecure)
+    allow(client.channel_manager).to receive(:get_channel).and_return(channel)
+    expect(client.users).to be_a(TbankGrpc::Services::UsersService)
+    expect(client.users).to equal(client.users)
+  end
+
+  it 'exposes #market_data as MarketDataService' do
+    channel = GRPC::Core::Channel.new('localhost:50051', {}, :this_channel_is_insecure)
+    allow(client.channel_manager).to receive(:get_channel).and_return(channel)
+    expect(client.market_data).to be_a(TbankGrpc::Services::MarketDataService)
+    expect(client.market_data).to equal(client.market_data)
+  end
+
   it 'exposes #helpers as Facade' do
     expect(client.helpers).to be_a(TbankGrpc::Helpers::Facade)
     expect(client.helpers.instruments).to be_a(TbankGrpc::Helpers::InstrumentsHelper)
+    expect(client.helpers.market_data).to be_a(TbankGrpc::Helpers::MarketDataHelper)
     expect(client.helpers).to equal(client.helpers)
   end